fix(inventario): existencia como entero, no decimal(12,3)
inventories.stock pasa de decimal(12,3) a unsignedInteger: el negocio
solo vende piezas completas, nunca fracciones, así que mostraba
"0.000" donde debía decir "0". sale_items.quantity (Fase 4) se queda
en decimal a propósito, por si algún producto futuro sí se vendiera
por metros o kilos — el cambio es solo para inventories.

Se edita la migración de la Fase 2 en vez de agregar una de alteración
(todavía no hay despliegue real, un solo ambiente local).

+++ app/Http/Requests/Admin/InventoryAdjustRequest.php
@@ -18,7 +18,7 @@ class InventoryAdjustRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'stock' => ['required', 'numeric', 'min:0'],
+            'stock' => ['required', 'integer', 'min:0'],
         ];
     }
 }
