+++ database/migrations/2026_08_26_055719_create_inventories_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
             $table->foreignId('branch_id')->constrained()->cascadeOnDelete();
-            $table->decimal('stock', 12, 3)->default(0);
+            $table->unsignedInteger('stock')->default(0);
             $table->string('physical_location')->nullable();
             $table->timestamps();
